Convert server code from Kotlin/Ktor to Java/Spring Boot
In the long-term, I'm not confident that Kotlin will replace Java on the
backend nor that Ktor will replace Spring Boot as the most popular
JVM language or framework, respectively. I don't really have any desire
to be on the bleeding edge of tech or go off the beaten path for this so
I'd like to keep with the industry standards so that it's easier for me
to work on over time.